add reassign deleted user's tasks
