+++ cmd/task_tracker/internal/app/get_tasks.go
@@ -1,0 +1,24 @@
+package app
+
+import (
+	"github.com/gin-gonic/gin"
+	"net/http"
+)
+
+func (a *App) GetUserTasks(context *gin.Context) {
+	userID, ok := context.Value("userID").(string)
+	if userID == "" || !ok {
+		context.JSON(http.StatusBadRequest, gin.H{"error": "no userID or userID is not string"})
+		context.Abort()
+		return
+	}
+
+	tasks, err := a.repo.GetUserTasks(userID)
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		context.Abort()
+		return
+	}
+
+	context.JSON(http.StatusOK, tasks)
+}
