+++ cmd/task_tracker/main.go
@@ -4,8 +4,10 @@ import (
 	"log"
 	"uber-popug/cmd/task_tracker/internal/api"
 	"uber-popug/cmd/task_tracker/internal/app"
+	"uber-popug/cmd/task_tracker/internal/handler"
 	"uber-popug/cmd/task_tracker/internal/repository"
 	"uber-popug/pkg/kafka/consumer"
+	"uber-popug/pkg/kafka/producer"
 )
 
 var (
@@ -18,25 +20,27 @@ func main() {
 		log.Fatal(err)
 	}
 
+	// producers
+	cudProducerConfig := producer.NewConfig(brokers, "tasks-stream", "tasks-service")
+	beProducerConfig := producer.NewConfig(brokers, "tasks", "tasks-service")
+
+	cudProducer, err := producer.NewProducer(cudProducerConfig)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	beProducer, err := producer.NewProducer(beProducerConfig)
+	if err != nil {
+		log.Fatal(err)
+	}
+
 	app := app.NewApp(repo, cudProducer, beProducer)
 
+	// users' events consumer
 	cudConsumerConfig := consumer.NewConfig(brokers, []string{"users-stream"}, "tasks-service")
 
 	c, err := consumer.New(cudConsumerConfig)
-	c.OnMessage(app.DeleteUserTasks)
-
-	//cudProducerConfig := producer.NewConfig(brokers, "tasks-stream", "tasks-service")
-	//beProducerConfig := producer.NewConfig(brokers, "tasks", "tasks-service")
-
-	//cudProducer, err := producer.NewProducer(cudProducerConfig)
-	//if err != nil {
-	//	log.Fatal(err)
-	//}
-	//
-	//beProducer, err := producer.NewProducer(beProducerConfig)
-	//if err != nil {
-	//	log.Fatal(err)
-	//}
+	c.OnMessage(handler.NewHandler(app).Handle)
 
 	// Initialize Router
 	router := api.NewApi(app)
